+++ app/scripts/controllers/network/network-controller.js
@@ -2,18 +2,13 @@ import { strict as assert } from 'assert';
 import EventEmitter from 'events';
 import { ComposedStore, ObservableStore } from '@metamask/obs-store';
 import { JsonRpcEngine } from 'json-rpc-engine';
-import {
-  providerFromEngine,
-  providerFromMiddleware,
-} from '@metamask/eth-json-rpc-middleware';
+import { providerFromEngine } from '@metamask/eth-json-rpc-middleware';
 import log from 'loglevel';
 import {
   createSwappableProxy,
   createEventEmitterProxy,
 } from 'swappable-obj-proxy';
 import EthQuery from 'eth-query';
-import createFilterMiddleware from 'eth-json-rpc-filters';
-import createSubscriptionManager from 'eth-json-rpc-filters/subscriptionManager';
 import { v4 as random } from 'uuid';
 import {
   INFURA_PROVIDER_TYPES,
@@ -460,26 +455,9 @@ export default class NetworkController extends EventEmitter {
   }
 
   _setNetworkClient({ networkMiddleware, blockTracker }) {
-    const networkProvider = providerFromMiddleware(networkMiddleware);
-    const filterMiddleware = createFilterMiddleware({
-      provider: networkProvider,
-      blockTracker,
-    });
-    const subscriptionManager = createSubscriptionManager({
-      provider: networkProvider,
-      blockTracker,
-    });
-
     const engine = new JsonRpcEngine();
-    subscriptionManager.events.on('notification', (message) =>
-      engine.emit('notification', message),
-    );
-    engine.push(filterMiddleware);
-    engine.push(subscriptionManager.middleware);
     engine.push(networkMiddleware);
-
     const provider = providerFromEngine(engine);
-
     this._setProviderAndBlockTracker({ provider, blockTracker });
   }
