+++ app/scripts/controllers/permissions/specifications.js
@@ -260,12 +260,10 @@ export const unrestrictedMethods = Object.freeze([
   'eth_signTypedData_v1',
   'eth_signTypedData_v3',
   'eth_signTypedData_v4',
-  'eth_subscribe',
   'eth_submitHashrate',
   'eth_submitWork',
   'eth_syncing',
   'eth_uninstallFilter',
-  'eth_unsubscribe',
   'metamask_getProviderState',
   'metamask_watchAsset',
   'net_listening',
